Restrict user actions

+++ dashboard/views.py
@@ -17,6 +17,7 @@ def explore(request):
 
 def roadmap_detail(request, pk):
     roadmap = get_object_or_404(Roadmap, pk=pk)
+    print(request.user == roadmap.user)
     context = {
         'roadmap': roadmap
     }
@@ -29,10 +30,13 @@ def roadmap_detail(request, pk):
 def roadmap_form(request):
     form = RoadmapForm()
     if request.method == 'POST':
-        form = RoadmapForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect('explore')
+        Roadmap.objects.create(
+            user=request.user,
+            title=request.POST.get('title'),
+            description=request.POST.get('description')
+        )
+
+        return redirect('explore')
     context = {'form': form,
                'msg': 'Roadmap'}
     return render(request, 'dashboard/roadmap_form.html', context)
@@ -44,6 +48,9 @@ def roadmap_update(request, pk):
     form = RoadmapForm(instance=roadmap)
 
     if request.method == 'POST':
+        form = EntityForm(request.POST)
+        if request.user != roadmap.user:
+            return HttpResponse("Not valid")
         form = RoadmapForm(request.POST, instance=roadmap)
         if form.is_valid():
             form.save()
@@ -59,6 +66,8 @@ def roadmap_update(request, pk):
 def roadmap_delete(request, pk):
     roadmap = Roadmap.objects.get(id=pk)
     if request.method == 'POST':
+        if request.user != roadmap.user:
+            return HttpResponse("Not valid")
         roadmap.delete()
         return redirect('explore')
     return render(request, 'dashboard/delete.html', {'obj': roadmap})
@@ -69,10 +78,15 @@ def roadmap_delete(request, pk):
 def entity_form(request):
     form = EntityForm()
     if request.method == 'POST':
-        form = EntityForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect('explore')
+        roadmap = Roadmap.objects.get(id = request.POST.get('roadmap'))
+        if request.user != roadmap.user:
+            return HttpResponse("Not valid")
+        Entity.objects.create(
+            roadmap=roadmap,
+            title = request.POST.get('title'),
+            entity_url= request.POST.get('entity_url')
+        )
+        return redirect('explore')
     context = {'form': form,
                'msg': 'Source'}
     return render(request, 'dashboard/roadmap_form.html', context)
@@ -84,6 +98,9 @@ def entity_update(request, pk):
     form = EntityForm(instance=entity)
 
     if request.method == 'POST':
+        if request.user != entity.roadmap.user:
+            return HttpResponse("Not valid")
+
         form = EntityForm(request.POST, instance=entity)
         if form.is_valid():
             form.save()
@@ -97,6 +114,8 @@ def entity_update(request, pk):
 def entity_delete(request, pk):
     entity = Entity.objects.get(id=pk)
     if request.method == 'POST':
+        if request.user != entity.roadmap.user:
+            return HttpResponse("Not valid")
         entity.delete()
         return redirect('explore')
     return render(request, 'dashboard/delete.html', {'obj': entity})
